:recycle: Splitting review pages from main page to make search simpler
